requirements.txt
pandas>=1.0.0
pyodbc>=4.0.30
python-dotenv>=0.15.0
sqlalchemy>=1.4.0
openpyxl>=3.0.0
